backend/requirements.txt
fastapi
uvicorn[standard]
psycopg[binary]
redis>=5.0.0
pydantic
pytest
httpx
google-auth
google-auth-httplib2
google-api-python-client
PyJWT
